requirements.txt
mysql-connector-python==8.0.23
protobuf==3.14.0
PyMySQL==1.0.2
six==1.15.0
SQLAlchemy==1.3.22
